Migrate formatAgeToDays over into utils/formatters
