+++ src/utils/formatters.js
@@ -188,6 +188,17 @@ export const formatAddressShort = (address) => {
   return `${city}, ${state} ${postalCode}`;
 };
 
+// Formats the numeric age input to a human readable string. Eg. 1.5 = 1 day, 2.5 = 2 days
+export const formatAgeToDays = (age) => {
+  if (age < 1) {
+    return 'Less than 1 day';
+  }
+  if (age >= 1 && age < 2) {
+    return '1 day';
+  }
+  return `${Math.floor(age)} days`;
+};
+
 // Format orders type (ex: PERMANENT_CHANGE_OF_STATION => Permanent change of station)
 export function formatOrderType(orderType) {
   return orderType
